Refactored some more code
