+++ src/main/java/net/kerosilas/imageviewer/tasks/SlideshowTask.java
@@ -1,6 +1,7 @@
-package net.kerosilas.imageviewer.model;
+package net.kerosilas.imageviewer.tasks;
 
 import javafx.concurrent.Task;
+import net.kerosilas.imageviewer.model.ImageManager;
 
 public class SlideshowTask extends Task<Void> {
